feat: Refactor to a google cloud function
