+++ package.json
@@ -19,6 +19,7 @@
     "node": "22.7.0"
   },
   "dependencies": {
+    "@google-cloud/functions-framework": "^3.4.2",
     "@octokit/core": "^6.1.2",
     "@octokit/graphql": "^8.1.1",
     "@octokit/webhooks": "^13.3.0",
